将系统所有id字段转化成long类型，完善商城的搜索和详情等，以及其他已知bug修复

+++ src/main/java/com/cn/leedane/controller/BaseController.java
@@ -649,7 +649,7 @@ public class BaseController {
 	 * 从message中解析user数据
 	 * @return
 	 */
-	protected int getUserIdFromShiro(){
+	protected long getUserIdFromShiro(){
 		UserBean user = getMustLoginUserFromShiro();
 		if(user == null)
 			throw new MustLoginException();
